% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_session_stats.R
\name{general_session_stats}
\alias{general_session_stats}
\title{drivingsimR: Automating the boring stuff, so you can calculate and later visualize what you care about.}
\usage{
general_session_stats(
  .data,
  group_var,
  lag_n = 10,
  sim = "drivesafety",
  fps_assumption = NA
)
}
\description{
drivingsimR: Automating the boring stuff, so you can calculate and later visualize what you care about.
}
\keyword{driving,}
\keyword{person}
\keyword{session}
\keyword{summary}
\keyword{summary,}
